requirements.txt
django>=4.0
django-ninja
mysqlclient==2.0.3